fix: don't test for explicit matrix type
If it's not a string or a list we will assume it is a matrix and use it as such. If `M` cannot be converted to a matrix, then Sage will tell the user.

I noticed this bug because some Sage methods return a _sparse_ matrix, which thus didn't match the type check.