fix: make output of cached methods immutable
If we don't do this, the user can modify the value of the cache. The majority of this commit is just docstrings changing from lists to tuples.

This affects the helper functions `Quiver.__all_harder_narasimhan_types` and `Quiver.__canonical_decomposition`, and thus all output of functions depending on these.

This fixes #36.